% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickEdit.R
\name{quickEdit}
\alias{quickEdit}
\title{saveVersion - An offline version control function

This function allows you to quickly edit data in your system's default editor for .csv or .txt files.}
\usage{
quickEdit(x)
}
\arguments{
\item{x}{The variable for your R object.}
}
\description{
saveVersion - An offline version control function

This function allows you to quickly edit data in your system's default editor for .csv or .txt files.
}
\examples{
quickEdit(df)
}
\keyword{editor}
